fixed implicit sign conversions

+++ src/problems/p1/src/p1.cpp
@@ -22,8 +22,8 @@ int p1_loop(const int upperBound)
 int p1_rawloop_vector(const std::vector<int>& v)
 {
     int sum = 0;
-    for (int i = 0; i < v.size(); ++i) {
-        sum += v[i];
+    for (int i = 0; i < static_cast<int>(v.size()); ++i) { // I know there won't be a bound problem for my use case
+        sum += v[static_cast<std::size_t>(i)];
     }
     return sum;
 }
@@ -58,7 +58,7 @@ int p1_partial_sum_formula(const int upperBound)
 
 std::vector<int> generateVector(const int upperBound)
 {
-    std::vector<int> v(upperBound);
+    std::vector<int> v(static_cast<size_t>(upperBound));
     std::iota(v.begin(), v.end(), 1);
     return v;
 }
